requirements.txt
openai>=1.30.0
tavily-python>=0.5.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
